Store variable that had the issue on exception
Closes #71

The BulkExpressionSolver provides an interface for customizing return
values when it throws an UnboundVariableError or a ZeroDivisionError
while trying to evaluate an expression. Right now there isn't a way to
know the name of the variable that we were attempting to assign the
evaluated expression to. This commit attaches a `recipient_variable`
to the exception. This requires creating a customized subclass of
ZeroDivisionError that belongs to Dentaku.